+++ src/PSCompression/Commands/CompressGzipArchiveCommand.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.IO.Compression;
 using System.Management.Automation;
+using PSCompression.Extensions;
 
 namespace PSCompression;
