+++ src/PSCompression/Commands/CompressZipArchiveCommand.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Management.Automation;
+using PSCompression.Extensions;
 
 namespace PSCompression;
