+++ src/PSCompression/Commands/ExpandGzipArchiveCommand.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.Management.Automation;
 using System.Text;
+using PSCompression.Extensions;
 
 namespace PSCompression;
