+++ src/PSCompression/Commands/ExpandZipEntryCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Management.Automation;
+using PSCompression.Extensions;
 
 namespace PSCompression;
