+++ src/PSCompression/Commands/GetZipEntryCommand.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO.Compression;
 using System.Linq;
 using System.Management.Automation;
+using PSCompression.Extensions;
 
 namespace PSCompression;
