+++ src/PSCompression/ExceptionHelpers.cs
@@ -2,6 +2,8 @@ using System;
 using System.IO;
 using System.IO.Compression;
 using System.Management.Automation;
+using PSCompression.Exceptions;
+using PSCompression.Extensions;
 
 namespace PSCompression;
