+++ src/PSCompression/Exceptions/DuplicatedEntryException.cs
@@ -1,6 +1,6 @@
 using System.IO;
 
-namespace PSCompression;
+namespace PSCompression.Exceptions;
 
 public sealed class DuplicatedEntryException : IOException
 {
@@ -13,7 +13,5 @@ public sealed class DuplicatedEntryException : IOException
     }
 
     internal static DuplicatedEntryException Create(string path, string source) =>
-        new DuplicatedEntryException(
-            $"An entry with path '{path}' already exists in '{source}'.",
-            path: path);
+        new($"An entry with path '{path}' already exists in '{source}'.", path);
 }
