+++ src/PSCompression/Extensions/PathExtensions.cs
@@ -4,12 +4,23 @@ using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 using System.Management.Automation;
+using System.Text.RegularExpressions;
 using Microsoft.PowerShell.Commands;
 
-namespace PSCompression;
+namespace PSCompression.Extensions;
 
 internal static class PathExtensions
 {
+    private static readonly Regex s_reNormalize = new(
+        @"(?:^[a-z]:)?[\\/]+|(?<![\\/])$",
+        RegexOptions.Compiled | RegexOptions.IgnoreCase);
+
+    private static readonly Regex s_reEntryDir = new(
+        @"[\\/]$",
+        RegexOptions.Compiled | RegexOptions.RightToLeft);
+
+    private const string _directorySeparator = "/";
+
     [ThreadStatic]
     private static List<string>? s_normalizedPaths;
 
@@ -94,6 +105,17 @@ internal static class PathExtensions
     internal static string GetParent(this string path) =>
         Path.GetDirectoryName(path);
 
-    internal static string GetLeaf(this string path) =>
-        Path.GetFileName(path);
+    internal static string NormalizeEntryPath(this string path) =>
+        s_reNormalize.Replace(path, _directorySeparator).TrimStart('/');
+
+    internal static string NormalizeFileEntryPath(this string path) =>
+        NormalizeEntryPath(path).TrimEnd('/');
+
+    internal static bool IsDirectoryPath(this string path) =>
+        s_reEntryDir.IsMatch(path);
+
+    public static string NormalizePath(this string path) =>
+        s_reEntryDir.IsMatch(path)
+        ? NormalizeEntryPath(path)
+        : NormalizeFileEntryPath(path);
 }
