+++ src/PSCompression/Extensions/ZipEntryExtensions.cs
@@ -2,43 +2,21 @@ using System.IO;
 using System.IO.Compression;
 using System.Text.RegularExpressions;
 
-namespace PSCompression;
+namespace PSCompression.Extensions;
 
 public static class ZipEntryExtensions
 {
-    private static readonly Regex s_reNormalize = new(
-        @"(?:^[a-z]:)?[\\/]+|(?<![\\/])$",
-        RegexOptions.Compiled | RegexOptions.IgnoreCase);
-
-    private static readonly Regex s_reEntryDir = new(
-        @"[\\/]$",
-        RegexOptions.Compiled | RegexOptions.RightToLeft);
-
     private static readonly Regex s_reGetDirName = new(
         @"[^/]+(?=/$)",
         RegexOptions.Compiled | RegexOptions.RightToLeft);
 
     private const string _directorySeparator = "/";
 
-    internal static string NormalizeEntryPath(this string path) =>
-        s_reNormalize.Replace(path, _directorySeparator).TrimStart('/');
-
-    internal static string NormalizeFileEntryPath(this string path) =>
-        NormalizeEntryPath(path).TrimEnd('/');
-
-    internal static bool IsDirectoryPath(this string path) =>
-        s_reEntryDir.IsMatch(path);
-
-    public static string NormalizePath(this string path) =>
-        s_reEntryDir.IsMatch(path)
-            ? NormalizeEntryPath(path)
-            : NormalizeFileEntryPath(path);
-
     internal static string RelativeTo(this DirectoryInfo directory, int length) =>
-        NormalizeEntryPath(directory.FullName.Substring(length) + _directorySeparator);
+        (directory.FullName.Substring(length) + _directorySeparator).NormalizeEntryPath();
 
     internal static string RelativeTo(this FileInfo directory, int length) =>
-        NormalizeFileEntryPath(directory.FullName.Substring(length));
+        directory.FullName.Substring(length).NormalizeFileEntryPath();
 
     internal static ZipArchiveEntry CreateEntryFromFile(
         this ZipArchive zip,
