+++ src/PSCompression/SortingOps.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using PSCompression.Extensions;
 
 namespace PSCompression;
