+++ src/PSCompression/ZipEntryBase.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO;
 using System.IO.Compression;
+using PSCompression.Extensions;
 
 namespace PSCompression;
