+++ src/PSCompression/ZipEntryDirectory.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO.Compression;
 using System.Linq;
+using PSCompression.Extensions;
 
 namespace PSCompression;
