+++ src/PSCompression/ZipEntryFile.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using System.IO.Compression;
+using PSCompression.Extensions;
 
 namespace PSCompression;
