+++ src/PSCompression/internal/_Format.cs
@@ -3,6 +3,7 @@ using System.ComponentModel;
 using System.Globalization;
 using System.IO;
 using System.Management.Automation;
+using PSCompression.Extensions;
 
 namespace PSCompression.Internal;
